// models/enums/OrganizationType.ts
export enum OrganizationType {
  NGO = 'NGO',                       // 非政府組織
  SOCIAL_ENTERPRISE = 'SOCIAL_ENTERPRISE', // 社會企業
  COOPERATIVE = 'COOPERATIVE',       // 合作社
  FOUNDATION = 'FOUNDATION',         // 基金會
  ASSOCIATION = 'ASSOCIATION',       // 協會
  EDUCATIONAL = 'EDUCATIONAL',       // 教育機構
  RELIGIOUS = 'RELIGIOUS',           // 宗教組織
  COMMUNITY = 'COMMUNITY',           // 社區組織
  GOVERNMENT = 'GOVERNMENT',         // 政府機構
  OTHER = 'OTHER'                    // 其他
}